prefixconf: add lease timings for prefixes
We want them to expire so we can actually scan them via
"ifconfig -L" but since they never expired before set
preferred lifetime to valid lifetime so they never deprecate.

Honor the no-release flag although this is a bit weird
in practise.  IA-NA is already using vltime/pltime but
their times are much closer and it's likely not good to
change that now since we are only interested in prefix
behaviour.