prefixconf: fix the re-adding during prefix updates
Since c69d026ab19d84e3 we no longer avoid re-adding the prefixes because
they expire.  This creates an issue since the code is meant to never update
and adds address duplicates to the existing tracked addresses.

To fix this reiterate over the existing prefix interface list, update
the timings and call the address function.